larger icon added, fix merge issues
